fastapi_app/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
numpy>=1.24.0
opencv-python-headless>=4.9.0
prometheus-client>=0.20.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
Pillow>=10.0.0
ultralytics
torch
torchvision
